LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/prefers-reduced-data-expected.txt


FAIL Should be known: '(prefers-reduced-data)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(prefers-reduced-data: no-preference)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(prefers-reduced-data: reduce)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: 0)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: none)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: 10px)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: no-preference reduce)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: reduced)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(prefers-reduced-data: no-preference/reduce)' assert_true: Can parse with JS expected true got false
FAIL Check that no-preference evaluates to false in the boolean context assert_equals: expected true but got false
PASS Check that invalid evaluates to false
